renumber

+++ renumber_models.py
@@ -6,10 +6,15 @@ fns = [fn for fn in os.listdir(ifp) if fn.startswith('id-')]
 fns.sort()
 
 idx = 0
-for fn in fns:
-    src = os.path.join(ifp, fn)
+while True:
     dst = os.path.join(ifp, 'id-{:08d}'.format(idx))
-    while os.path.exists(dst):
+    if os.path.exists(dst):
         idx += 1
-        dst = os.path.join(ifp, 'id-{:08d}'.format(idx))
+        continue
+
+    src = os.path.join(ifp, fns[-1])
     os.rename(src, dst)
+    del fns[-1]
+    idx += 1
+    if len(fns) == 0:
+        break
